fix share menu e2e test after rebasing on develop
